+++ link-move/src/main/java/com/nhl/link/move/runtime/extractor/IExtractorService.java
@@ -5,9 +5,9 @@ import com.nhl.link.move.extractor.model.ExtractorName;
 
 public interface IExtractorService {
 
-	/**
-	 * Returns an {@link Extractor} identified by {@link ExtractorName} or
-	 * throws an exception if such an extractor is not available.
-	 */
-	Extractor getExtractor(ExtractorName name);
+    /**
+     * Returns an {@link Extractor} identified by {@link ExtractorName} or throws an exception if such an extractor
+     * is not available.
+     */
+    Extractor getExtractor(ExtractorName name);
 }